+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - apply `box-sizing: border-box` to all items
 - left and maxWidth styles for CarouselModal to work with `box-sizing: border-box` change.
+- when autoplay is `false` for a video item, seeking via progress bar no longer plays the video.
+- the progress bar now tracks whether the video was playing before a seek occurs and plays the video onMouseUp accordingly
 
 ### Removed